requirements.txt
Flask==2.0.2
uvicorn>=0.15.0,<0.16.0
gunicorn==20.1.0
azure-storage-blob==12.20.0
elasticsearch==8.4.3
geopy==2.3.0
requests==2.31.0
PyPDF2==3.0.1
openai==1.30.3
pandas==2.2.2
doc2text==0.2.4
docx2txt==0.8
bs4==0.0.2
aiohttp==3.9.5
lxml==5.2.2